feat(navbar): packages look
