+++ src/app/shared/navbar/navbar.component.scss
@@ -18,13 +18,11 @@
 }
 
 .navbar__button--active {
-  border: 1px solid rgb(214, 184, 214) !important;
   background-color: rgb(214, 184, 214);
   color: white;
 }
 
 .navbar__button--inactive {
-  border: 1px solid black;
   background-color: white;
   color: black;
 }
