+++ src/app/shared/side-card-packages/SideCardPackage.ts
@@ -1,0 +1,5 @@
+import { PackagesBox } from './side-card-packages-box/side-card-packages-box.component';
+
+export class SideCadrPackage {
+  constructor(public box: PackagesBox, public descriptions: string[]) {}
+}
