+++ src/app/shared/side-card-packages/side-card-packages-box/side-card-packages-box.component.html
@@ -1,32 +1,39 @@
-<div fxLayout="column" fxLayoutAlign="start center" fxLayoutGap="20px">
+<div fxLayout="column" fxLayoutAlign="start" fxLayoutGap="5%">
   <div
     *ngIf="this.package"
     class="sc-packages-box__container"
     fxLayout="column"
-    fxLayoutAlign="center center"
+    fxLayoutAlign="space-around center"
     fxLayoutGap="10px"
   >
     <span class="packages-box__type">
       {{ this.package?.box?.type }}
     </span>
-    <span class="packages-box__price">
-      ${{ this.package?.box?.price }}
-    </span>
+    <span class="packages-box__price"> ${{ this.package?.box?.price }} </span>
     <span class="packages-box__description">
       {{ this.package?.box?.description }}
     </span>
 
-    <button (click)="this.selectPackage()">
+    <div
+      fxLayoutAlign="center center"
+      class="packages-box__select"
+      (click)="this.selectPackage()"
+    >
       SELECT
-    </button>
+    </div>
   </div>
 
-  <div fxLayout="column" fxLayoutAlign="start center" fxLayoutGap="10px">
+  <div
+    class="packages-out-box__conatainer"
+    fxLayout="column"
+    fxLayoutAlign="start center"
+    fxLayoutGap="10px"
+  >
     <div
       *ngFor="let desc of this.package.descriptions"
-      class="packages-box__descriptions"
+      class="packages-out-box__description"
     >
-      <small>{{ desc }}</small>
+      {{ desc }}
       <br />
     </div>
   </div>
