+++ src/app/shared/side-card-packages/side-card-packages-box/side-card-packages-box.component.scss
@@ -1,7 +1,56 @@
+@import "../../../../styles-colors.scss";
+
+$sc-packages-width: 184px;
+
 .sc-packages-box__container {
-  width: 160px;
-  height: 190px;
+  width: $sc-packages-width;
+  height: 60%;
   border: 1px solid white;
 
   text-align: center;
+}
+
+.packages-box__type {
+  margin-top: 10%;
+  font-size: 1.15rem;
+  font-weight: 400;
+}
+
+.packages-box__price {
+  font-size: 2rem;
+  font-weight: 700;
+
+  text-shadow: 4px 4px 5px $text-shadow;
+}
+
+.packages-box__description {
+  font-size: 0.85rem;
+  font-weight: 300;
+  width: 90%;
+}
+
+.packages-box__select {
+  width: 90%;
+
+  margin-bottom: 5%;
+  padding: 3.5%;
+
+  font-size: 0.8rem;
+  font-weight: 400;
+
+  border: 1px solid white;
+  cursor: pointer !important;
+}
+
+.packages-out-box__conatainer {
+  width: $sc-packages-width;
+  padding: 2.5%;
+}
+
+.packages-out-box__description {
+  font-size: 0.6rem;
+  font-weight: 400;
+
+  text-align: center;
+  white-space: pre-line;
 }
