+++ src/app/shared/side-card-packages/side-card-packages.component.html
@@ -1,7 +1,7 @@
 <app-side-card>
   <div fxLayout="row" fxLayoutAlign="center center" fxLayoutGap="10px">
-    <div fxFlex="10"></div>
-    <div fxFlex>
+    <div fxFlex="17"></div>
+    <div fxLayout="column" fxLayoutAlign="start start" fxLayoutGap="3%" fxFlex>
       <div
         class="sd-packages__header"
         fxLayout="column"
@@ -14,7 +14,7 @@
         <span class="sd-packages__subtitle">
           HOW BIG IS YOUR SPACE?
         </span>
-        <span class="sd-packages__title">
+        <span class="sd-packages__subtitle-decs">
           We specialise in apartments and lofts renovation projects and our core
           packages are based on the size of your space.
         </span>
