+++ src/app/shared/side-card-packages/side-card-packages.component.ts
@@ -1,75 +1,6 @@
 import { Component, OnInit } from '@angular/core';
-import {
-  PackagesBox,
-  PackageType,
-} from './side-card-packages-box/side-card-packages-box.component';
-
-export class SideCadrPackage {
-  constructor(public box: PackagesBox, public descriptions: string[]) {}
-}
-
-const PACKAGES = [
-  new SideCadrPackage(
-    new PackagesBox(
-      PackageType.SMALL,
-      399,
-      `For small spaces up to 40SQM/430SQFT`
-    ),
-    [
-      `S package includes one best solution
-    based on the space possibilities
-    and your needs`,
-      `INCLUDES
-    - New redesigned floor plan
-    - Furniture disposition plan
-    - Redesigned floor plan
-    with detailed measurements
-    - Demolition and deconstruction plan
-    - Project description`,
-      `In total you will recieve 5 different files.`,
-    ]
-  ),
-  new SideCadrPackage(
-    new PackagesBox(
-      PackageType.MEDIUM,
-      599,
-      `For apartment up to 75 square meters`
-    ),
-    [
-      `S package includes one best solution
-    based on the space possibilities
-    and your needs`,
-      `INCLUDES
-    - New redesigned floor plan
-    - Furniture disposition plan
-    - Redesigned floor plan
-    with detailed measurements
-    - Demolition and deconstruction plan
-    - Project description`,
-      `In total you will recieve 5 different files.`,
-    ]
-  ),
-  new SideCadrPackage(
-    new PackagesBox(
-      PackageType.LARGE,
-      799,
-      `For apartment up to 100 square meters`
-    ),
-    [
-      `S package includes one best solution
-    based on the space possibilities
-    and your needs`,
-      `INCLUDES
-    - New redesigned floor plan
-    - Furniture disposition plan
-    - Redesigned floor plan
-    with detailed measurements
-    - Demolition and deconstruction plan
-    - Project description`,
-      `In total you will recieve 5 different files.`,
-    ]
-  ),
-];
+import { PACKAGES } from './side-card-packages.content';
+import { SideCadrPackage } from './SideCardPackage';
 
 @Component({
   selector: 'app-side-card-packages',
