+++ src/styles-colors.scss
@@ -1,4 +1,5 @@
 $white: #ffffff;
+$text-shadow: rgb(165, 165, 165);
 
 .color--white {
   color: $white;
